feat(workflow): refactor session components and enhance time formatting utility
